backend/requirements.txt
pywebview==6.0
fastapi==0.115.12
pyinstaller==6.12.0
uvicorn==0.34.0
httpx==0.28.1
ruff
python-dotenv==1.2.2
pyperclip==1.11.0
pillow==11.3.0